+++ composer.json
@@ -2,7 +2,7 @@
 	"name": "connectpos/module-performance",
 	"description": "N/A",
 	"type": "magento2-module",
-	"version": "21.05.06",
+	"version": "21.05.20",
 	"license": [
 		"proprietary"
 	],
